+++ src/App.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { ref, onUnmounted } from 'vue'
-import type { TimerState, Task } from './types'
+import type { TimerState } from './types'
 import TaskList from './components/TaskList.vue'
 
 const timerState = ref<TimerState>('idle')
@@ -8,42 +8,6 @@ const Time = 60 * 25
 const time = ref(Time)
 const intervalId = ref<ReturnType<typeof setInterval>>()
 const isTaskListVisible = ref(false)
-
-const getStoredTaskList = (): Task[] => {
-  try {
-    const stored = localStorage.getItem('taskList')
-    if (!stored) return []
-    return JSON.parse(stored)
-  } catch (error) {
-    console.error(error)
-    return []
-  }
-}
-const taskList = ref<Task[]>(getStoredTaskList())
-
-const addTask = (newTaskContent: string) => {
-  const newTask = {
-    id: Date.now(),
-    content: newTaskContent,
-    isCompleted: false,
-  }
-  taskList.value.push(newTask)
-  try {
-    localStorage.setItem('taskList', JSON.stringify(taskList.value))
-  } catch (error) {
-    console.error(error)
-  }
-}
-
-const updateTaskOrder = (newTaskList: Task[]) => {
-  console.log(newTaskList)
-  taskList.value = newTaskList
-  try {
-    localStorage.setItem('taskList', JSON.stringify(taskList.value))
-  } catch (error) {
-    console.error(error)
-  }
-}
 
 const startTimer = () => {
   timerState.value = 'running'
@@ -96,11 +60,6 @@ onUnmounted(() => {
       ></Button>
       <Button v-if="timerState == 'finished'" label="Restart" @click="reStartTimer"></Button>
     </div>
-    <TaskList
-      v-model:visible="isTaskListVisible"
-      v-model:task-list="taskList"
-      @add-task="addTask"
-      @update-task-order="updateTaskOrder"
-    />
+    <TaskList v-model:visible="isTaskListVisible" />
   </div>
 </template>
